Fix NullPointerException in example app
Fixed NullPointerException that was being thrown on line 93, caused by
a method call in the Hawk library before Hawk has finished initializing.
This was fixed by moving the method calls to be executed after Hawk has
been initialized into the onSuccess method of the HawkBuilder.Callback
to ensure that Hawk will always be initialized before any method calls
utilizing it are made.